Use ImportCsvCommand `to_cache` feature
