+++ perfil/election/management/commands/load_assets.py
@@ -1,8 +1,6 @@
-from datetime import datetime
 from textwrap import dedent
 
-from django.core.cache import cache
-
+from perfil.election.management.commands import election_keys
 from perfil.election.models import Election, Asset
 from perfil.utils.management.commands import ImportCsvCommand
 
@@ -15,27 +13,11 @@ class Command(ImportCsvCommand):
 
     model = Asset
     bulk_size = 2 ** 13
-
-    @staticmethod
-    def election_key(election):
-        if isinstance(election, Election):
-            keys = (str(election.year), election.state,
-                    election.candidate_sequential)
-        else:
-            keys = (str(election['ano_eleicao']), election['sigla_uf'],
-                    election['sq_candidato'])
-
-        key = '-'.join(keys)
-        return f'election-{key}'
+    to_cache = (Election, election_keys),
 
     def serialize(self, reader):
-        print('Caching elections data…')
-        for election in Election.objects.all().iterator():
-            cache.set(self.election_key(election), election.id, 60 * 60 * 4)
-
-        print('Import asset data…')
         for line in reader:
-            election_id = cache.get(self.election_key(line))
+            election_id = self.cache.get(election_keys(line))
             if election_id:
                 yield Asset(
                     election_id=election_id,
